requirements.txt
Pillow==11.0.0
pyzbar==0.1.9
qrcode==8.0
